Extract `actions` to separate submodule
Break some pieces apart for better separation.
