+++ children/mutations.js
@@ -1,47 +1,6 @@
 import R from 'ramda';
 import { stream } from 'kefir';
-
-/**
- * Node added action constant.
- *
- * @type {string}
- */
-export const NODE_ADDED = 'NODE_ADDED';
-
-/**
- * Creates a new NODE_ADDED action.
- *
- * @param {Node} target - Mutation target.
- * @param {Node} node - Node added.
- * @returns {Action} Node added action.
- */
-export const nodeAdded = R.curry(function nodeAdded(target, node) {
-    return {
-        type: NODE_ADDED,
-        payload: { target, node }
-    };
-});
-
-/**
- * Node removed action constant.
- *
- * @type {string}
- */
-export const NODE_REMOVED = 'NODE_REMOVED';
-
-/**
- * Creates a new NODE_REMOVED action.
- *
- * @param {Node} target - Mutation target.
- * @param {Node} node - Node removed.
- * @returns {Action} Node removed action.
- */
-export const nodeRemoved = R.curry(function nodeRemoved(target, node) {
-    return {
-        type: NODE_REMOVED,
-        payload: { target, node }
-    };
-});
+import { nodeAdded, nodeRemoved } from './actions';
 
 /**
  * Determines whether the node is relevant to stream consumers.
